Use int, not short for events in the event API.
This fixes some -Wconversion warnings and fixes an inconsistency
between the libsudo_util event API and the plugin event API.  The
actual struct internals still use shorts to avoid changing the ABI.

+++ lib/util/event_poll.c
@@ -120,7 +120,7 @@ sudo_ev_add_impl(struct sudo_event_base *base, struct sudo_event *ev)
     /* Fill in pfd entry. */
     sudo_debug_printf(SUDO_DEBUG_DEBUG|SUDO_DEBUG_LINENO,
 	"%s: choosing free slot %d", __func__, base->pfd_free);
-    ev->pfd_idx = base->pfd_free;
+    ev->pfd_idx = (short)base->pfd_free;
     pfd = &base->pfds[ev->pfd_idx];
     pfd->fd = ev->fd;
     pfd->events = 0;
@@ -226,7 +226,7 @@ sudo_ev_scan_impl(struct sudo_event_base *base, unsigned int flags)
 		sudo_debug_printf(SUDO_DEBUG_DEBUG,
 		    "%s: polled fd %d, events %d, activating %p",
 		    __func__, ev->fd, what, ev);
-		ev->revents = what;
+		ev->revents = (short)what;
 		sudo_ev_activate(base, ev);
 	    }
 	}
